app/requirements.txt
sqlalchemy
pandas
numpy
pydantic
python-dotenv
psycopg2